Create ubuntuinstaller.sh
Proposing this script to be hosted here because I was tired of dealing with copying and pasting commands all the time. This should assist first time setup for users with Ubuntu. In the future this script can be modified to detect the current system and install based on the OS.